+++ cognite/client/_api_client.py
@@ -42,6 +42,7 @@ from cognite.client.data_classes._base import (
     T_CogniteResource,
     T_CogniteResourceList,
 )
+from cognite.client.data_classes.filters import Filter
 from cognite.client.exceptions import CogniteAPIError, CogniteNotFoundError
 from cognite.client.utils._auxiliary import is_unlimited, split_into_chunks
 from cognite.client.utils._concurrency import TaskExecutor
@@ -368,11 +369,12 @@ class APIClient:
         limit: Optional[int] = None,
         chunk_size: Optional[int] = None,
         filter: Optional[Dict[str, Any]] = None,
-        sort: Optional[Sequence[str]] = None,
+        sort: Optional[Sequence[str | dict]] = None,
         other_params: Optional[Dict[str, Any]] = None,
         partitions: Optional[int] = None,
         headers: Optional[Dict[str, Any]] = None,
         initial_cursor: Optional[str] = None,
+        advanced_filter: Optional[dict | Filter] = None,
     ) -> Union[Iterator[T_CogniteResourceList], Iterator[T_CogniteResource]]:
         if is_unlimited(limit):
             limit = None
@@ -419,6 +421,12 @@ class APIClient:
                     body: dict[str, Any] = {"limit": current_limit, "cursor": next_cursor, **(other_params or {})}
                     if filter:
                         body["filter"] = filter
+                    if advanced_filter:
+                        body["advancedFilter"] = (
+                            advanced_filter.dump(camel_case_property=True)
+                            if isinstance(advanced_filter, Filter)
+                            else advanced_filter
+                        )
                     if sort is not None:
                         body["sort"] = sort
                     res = self._post(url_path=url_path or resource_path + "/list", json=body, headers=headers)
@@ -494,9 +502,10 @@ class APIClient:
         filter: Optional[Dict] = None,
         other_params: Optional[Dict] = None,
         partitions: Optional[int] = None,
-        sort: Optional[Sequence[str]] = None,
+        sort: Optional[Sequence[str | dict]] = None,
         headers: Optional[Dict] = None,
         initial_cursor: Optional[str] = None,
+        advanced_filter: Optional[dict | Filter] = None,
     ) -> T_CogniteResourceList:
         if partitions:
             if not is_unlimited(limit):
@@ -528,6 +537,7 @@ class APIClient:
             other_params=other_params,
             headers=headers,
             initial_cursor=initial_cursor,
+            advanced_filter=advanced_filter,
         ):
             items.extend(resource_list.data)
         return list_cls(items, cognite_client=self._cognite_client)
@@ -541,6 +551,7 @@ class APIClient:
         filter: Optional[Dict] = None,
         other_params: Optional[Dict] = None,
         headers: Optional[Dict] = None,
+        advanced_filter: Optional[dict | Filter] = None,
     ) -> T_CogniteResourceList:
         def get_partition(partition: int) -> List[Dict[str, Any]]:
             next_cursor = None
@@ -554,6 +565,12 @@ class APIClient:
                         "partition": partition,
                         **(other_params or {}),
                     }
+                    if advanced_filter:
+                        body["advancedFilter"] = (
+                            advanced_filter.dump(camel_case_property=True)
+                            if isinstance(advanced_filter, Filter)
+                            else advanced_filter
+                        )
                     res = self._post(
                         url_path=(resource_path or self._RESOURCE_PATH) + "/list", json=body, headers=headers
                     )